+++ dynamicscrmapi/src/test/java/MicrosoftDataAccessObjectTestSuite.java
@@ -1,0 +1,12 @@
+import org.junit.runner.RunWith;
+import org.junit.runners.Suite;
+
+@RunWith(Suite.class)
+
+@Suite.SuiteClasses({
+        TestJunit.class,
+        DynamicsDaoTest.class
+})
+
+public class MicrosoftDataAccessObjectTestSuite { }
+
